requirement.txt
python-dotenv
pandas 
numpy 
pymongo[srv] == 4.7
pymongo
certifi
scikit-learn
joblib
scipy
matplotlib
seaborn
## -e .